+++ CHANGES.md
@@ -1,6 +1,7 @@
 0.1.1
 -----
 
+* Initial signaling support via MyActor#signal and MyActor#wait
 * Just "include Celluloid" works in lieu of "include Celluloid::Actor"
 * Futures terminate implicitly when their values are obtained
 * Add an underscore prefix to all of Celluloid's instance variables so they don't